requirements.txt
selenium==4.18.1
python-dotenv==1.0.1
requests==2.31.0
logging==0.4.9.6
rich==13.7.0
pyyaml>=6.0 